app/report/views.py
from rest_framework import permissions, viewsets

from app.report.models import Report
from app.report.serializers import ReportSerializer


class ReportViewSet(viewsets.ModelViewSet):
    queryset = Report.objects.all()
    serializer_class = ReportSerializer
    permission_classes = [permissions.IsAuthenticated]
