requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.7.0
click==8.1.6
coverage==7.3.0
decorator==5.1.1
dj-database-url==2.1.0
dj-static==0.0.6
Django==5.0.7
django-auditlog==3.0.0
django-debug-toolbar==4.2.0
django-filter==24.2
django-object-actions==4.1.0
django-test-without-migrations==0.6
djangorestframework==3.15.2
drf-yasg==1.21.7
executing==1.2.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.14.0
jedi==0.19.0
Markdown==3.6
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.4.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pytest==8.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
setuptools==71.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
static3==0.7.0
traitlets==5.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
wcwidth==0.2.6
